pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coral-tpu-mcp"
version = "0.1.0"
description = "Coral TPU MCP Server - Fast local ML inference for agentic systems"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Marc", email = "user@example.com"}
]
keywords = ["coral", "tpu", "mcp", "edge-ml", "inference"]

dependencies = [
    "mcp>=1.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    # pycoral and tflite-runtime installed separately in coral-venv
]

[project.optional-dependencies]
text = [
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short -p no:anyio"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/coral_tpu_mcp"]
branch = true
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true

[project.scripts]
coral-tpu-mcp = "coral_tpu_mcp.server:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt"]
